requirements.txt
requests==2.32.5
psycopg2-binary==2.9.11
python-dotenv==1.2.2
yfinance==1.2.0
pytrends==4.9.2
